refactor: Modify product class into enums
