+++ src/main/java/domain/Underwriting.java
@@ -1,6 +1,5 @@
 package domain;
 
-import domain.product.Product;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
