--- src/main/java/domain/product/Automobile.java
@@ -1,8 +1,0 @@
-package domain.product;
-
-public class Automobile extends Product {
-
-  public Automobile() {
-
-  }
-}
