--- src/main/java/domain/product/Driver.java
@@ -1,8 +1,0 @@
-package domain.product;
-
-public class Driver extends Product {
-
-  public Driver() {
-
-  }
-}
