--- src/main/java/domain/product/Fire.java
@@ -1,8 +1,0 @@
-package domain.product;
-
-public class Fire extends Product {
-
-  public Fire() {
-
-  }
-}
