--- src/main/java/domain/product/Travel.java
@@ -1,8 +1,0 @@
-package domain.product;
-
-public class Travel extends Product {
-
-  public Travel() {
-
-  }
-}
