+++ src/main/java/dto/ContractServiceDto.java
@@ -1,7 +1,7 @@
 package dto;
 
 import domain.Customer;
-import domain.product.Product;
+import domain.Product;
 import java.io.File;
 import lombok.Builder;
 import lombok.Data;
