+++ src/main/java/enums/ProductType.java
@@ -1,0 +1,8 @@
+package enums;
+
+public enum ProductType {
+  AUTOMOBILE_Insurance,
+  DRIVER_Insurance,
+  FIRE_Insurance,
+  TRAVEL_Insurance
+}
